requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
chromadb
qdrant-client
asyncpg
sqlalchemy[asyncio]
alembic
greenlet
tiktoken
slowapi
pydantic>=2.0
email-validator
